clean up workspace conditions
