+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/conditions/constants.ts
@@ -1,0 +1,3 @@
+export * from './workspace-alias/constants.js';
+export * from './workspace-entity-is-new/constants.js';
+export * from './workspace-entity-type/constants.js';
