+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/conditions/manifests.ts
@@ -1,9 +1,9 @@
-import { manifest as workspaceAliasCondition } from './workspace-alias.condition.js';
-import { manifest as workspaceEntityTypeCondition } from './workspace-entity-type.condition.js';
-import { manifest as workspaceEntityIsNewCondition } from './workspace-entity-is-new.condition.js';
+import { manifests as workspaceAliasCondition } from './workspace-alias/manifests.js';
+import { manifests as workspaceEntityIsNewCondition } from './workspace-entity-is-new/manifests.js';
+import { manifests as workspaceEntityTypeCondition } from './workspace-entity-type/manifests.js';
 
 export const manifests: Array<UmbExtensionManifest> = [
-	workspaceEntityIsNewCondition,
-	workspaceAliasCondition,
-	workspaceEntityTypeCondition,
+	...workspaceAliasCondition,
+	...workspaceEntityIsNewCondition,
+	...workspaceEntityTypeCondition,
 ];
