+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/conditions/workspace-alias/constants.ts
@@ -1,0 +1,4 @@
+/**
+ * Workspace alias condition alias
+ */
+export const UMB_WORKSPACE_CONDITION_ALIAS = 'Umb.Condition.WorkspaceAlias';
