+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/conditions/workspace-alias/workspace-alias.condition.ts
@@ -1,7 +1,7 @@
-import { UMB_WORKSPACE_CONTEXT } from '../workspace.context-token.js';
-import type { UmbWorkspaceContext } from '../workspace-context.interface.js';
-import type { WorkspaceAliasConditionConfig } from './types.js';
-import { UMB_WORKSPACE_CONDITION_ALIAS } from './const.js';
+import type { WorkspaceAliasConditionConfig } from '../types.js';
+import type { UmbWorkspaceContext } from '../../workspace-context.interface.js';
+import { UMB_WORKSPACE_CONTEXT } from '../../workspace.context-token.js';
+import { UMB_WORKSPACE_CONDITION_ALIAS } from './constants.js';
 import { UmbConditionBase } from '@umbraco-cms/backoffice/extension-registry';
 import type { UmbConditionControllerArguments, UmbExtensionCondition } from '@umbraco-cms/backoffice/extension-api';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
@@ -35,10 +35,3 @@ export class UmbWorkspaceAliasCondition
 		}
 	}
 }
-
-export const manifest: UmbExtensionManifest = {
-	type: 'condition',
-	name: 'Workspace Alias Condition',
-	alias: UMB_WORKSPACE_CONDITION_ALIAS,
-	api: UmbWorkspaceAliasCondition,
-};
