+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/conditions/workspace-entity-is-new/constants.ts
@@ -2,8 +2,3 @@
  * Workspace entity is new condition alias
  */
 export const UMB_WORKSPACE_ENTITY_IS_NEW_CONDITION_ALIAS = 'Umb.Condition.WorkspaceEntityIsNew';
-
-/**
- * Workspace alias condition alias
- */
-export const UMB_WORKSPACE_CONDITION_ALIAS = 'Umb.Condition.WorkspaceAlias';
