+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/conditions/workspace-entity-type/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_WORKSPACE_ENTITY_TYPE_CONDITION_ALIAS = 'Umb.Condition.WorkspaceEntityType';
