+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/conditions/workspace-entity-type/types.ts
@@ -1,0 +1,23 @@
+import type { UMB_WORKSPACE_ENTITY_TYPE_CONDITION_ALIAS } from './constants.js';
+import type { UmbConditionConfigBase } from '@umbraco-cms/backoffice/extension-api';
+
+export type UmbWorkspaceEntityTypeConditionConfig = UmbConditionConfigBase<
+	typeof UMB_WORKSPACE_ENTITY_TYPE_CONDITION_ALIAS
+> & {
+	/**
+	 * Define the workspace that this extension should be available in
+	 * @example
+	 * "Document"
+	 */
+	match: string;
+};
+/**
+ * @deprecated Use `UmbWorkspaceEntityTypeConditionConfig` instead. This will be removed in Umbraco 17.
+ */
+export type WorkspaceEntityTypeConditionConfig = UmbWorkspaceEntityTypeConditionConfig;
+
+declare global {
+	interface UmbExtensionConditionConfigMap {
+		umbWorkspaceEntityType: UmbWorkspaceEntityTypeConditionConfig;
+	}
+}
