+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/conditions/workspace-entity-type/workspace-entity-type.condition.ts
@@ -1,5 +1,5 @@
-import { UMB_WORKSPACE_CONTEXT } from '../workspace.context-token.js';
-import type { UmbWorkspaceEntityTypeConditionConfig } from './types.js';
+import { UMB_WORKSPACE_CONTEXT } from '../../workspace.context-token.js';
+import type { UmbWorkspaceEntityTypeConditionConfig } from '../types.js';
 import { UmbConditionBase } from '@umbraco-cms/backoffice/extension-registry';
 import type { UmbConditionControllerArguments, UmbExtensionCondition } from '@umbraco-cms/backoffice/extension-api';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
@@ -15,10 +15,3 @@ export class UmbWorkspaceEntityTypeCondition
 		});
 	}
 }
-
-export const manifest: UmbExtensionManifest = {
-	type: 'condition',
-	name: 'Workspace Entity Type Condition',
-	alias: 'Umb.Condition.WorkspaceEntityType',
-	api: UmbWorkspaceEntityTypeCondition,
-};
