+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/constants.ts
@@ -1,2 +1,3 @@
-export { UMB_NAMABLE_WORKSPACE_CONTEXT } from './namable/index.js';
 export { UMB_ENTITY_DETAIL_WORKSPACE_CONTEXT } from './entity-detail/index.js';
+export { UMB_NAMABLE_WORKSPACE_CONTEXT } from './namable/index.js';
+export * from './conditions/constants.js';
