+++ src/Umbraco.Web.UI.Client/src/packages/core/workspace/index.ts
@@ -1,5 +1,4 @@
 export * from './components/index.js';
-export * from './conditions/const.js';
 export * from './constants.js';
 export * from './contexts/index.js';
 export * from './controllers/index.js';
